refactor(init): extract answer types into src/init/types.ts
Move InitAnswers and friends out of wizard.tsx (UI) so render.ts (pure)
no longer imports from a UI module. No shape changes.
