+++ src/init/render.ts
@@ -1,4 +1,4 @@
-import type { InitAnswers } from "./wizard.js";
+import type { InitAnswers } from "./types.js";
 
 const HEADER = `# ~/.isotopes/isotopes.yaml
 # Generated by \`isotopes init\`. See isotopes.example.yaml in the repo
