+++ src/init/wizard.tsx
@@ -8,38 +8,14 @@ import { Box, Text, useApp, useInput } from "ink";
 import SelectInput from "ink-select-input";
 import TextInput from "ink-text-input";
 
-// ---------------------------------------------------------------------------
-// Types
-// ---------------------------------------------------------------------------
-
-export type LlmChoice = "ghc-proxy" | "skip";
-export type ChannelChoice = "discord" | "skip";
-export type CodingAgentChoice = "claude" | "skip";
-
-export interface GhcProxyAnswers {
-  baseUrl: string;
-  apiKey: string;
-  model: string;
-}
-
-export type DmPolicyChoice = "disabled" | "allowlist";
-export type GroupPolicyChoice = "disabled" | "allowlist" | "open";
-
-export interface DiscordAnswers {
-  token: string;
-  dmPolicy: DmPolicyChoice;
-  dmUserId?: string;
-  groupPolicy: GroupPolicyChoice;
-  groupAllowlist?: string[];
-}
-
-export interface InitAnswers {
-  llm: LlmChoice;
-  ghcProxy?: GhcProxyAnswers;
-  channel: ChannelChoice;
-  discord?: DiscordAnswers;
-  codingAgent: CodingAgentChoice;
-}
+import type {
+  ChannelChoice,
+  CodingAgentChoice,
+  DmPolicyChoice,
+  GroupPolicyChoice,
+  InitAnswers,
+  LlmChoice,
+} from "./types.js";
 
 // ---------------------------------------------------------------------------
 // Defaults
